Initial implementation of a TODO application for the TUI
This version is functional and acceptable but missing a few
things such as a database interface and database migrations
for when upgrading the application to a newer version.
While there is an in-memory database added for the application
it is non-functional as the database cannot be stored over
multiple executions of the app.
The setup command assumes that direnv is installed on the
machine. This should be fixed.