Add database migration support
When a new database is loaded it will read the version of the
header and attempt to migrate the database.
The migration will always take the full path from the file version
up to the latest version of the database. This will ensure that
a user can always migrate from any version of todotui.
Todotui cannot yet handle the situation where the header is removed
from a file with todos in it.
A backup is not created when doing migrations and should be fixed.

Also did some refactoring of the __main__ so that it is not as
cluttered.